consolidate CcompilerAnnotator
